fix: minor fixes for ios c-interop, gradle build logic, and lint ci (#40)
* fix(ios): resolve type mismatch in CFNumberCreate

* build: clean up gradle scripts and lint configuration

* ci(lint): update android lint task and sarif path

+++ kpdfium/src/iosMain/kotlin/com/sorrowblue/kpdfium/IosPdfPage.kt
@@ -10,7 +10,6 @@ import kotlinx.coroutines.IO
 import kotlinx.coroutines.sync.withLock
 import kotlinx.coroutines.withContext
 import kotlinx.io.Sink
-import kotlinx.io.write
 import platform.CoreFoundation.*
 import platform.CoreGraphics.*
 import platform.ImageIO.*
@@ -154,7 +153,7 @@ internal class IosPdfPage(
             qualityVar.value = quality / QUALITY_PERCENT
             val qualityNum = CFNumberCreate(
                 null,
-                kCFNumberFloatType.toLong(),
+                kCFNumberFloatType,
                 qualityVar.ptr
             )
             if (qualityNum != null) {
